FleetRepReactor in progress

+++ engine/SegmentReactor.cpp
@@ -43,7 +43,7 @@ namespace Shipping {
 		int numPackages = _ptr->numPackages().value();
 		int capacity = segment_->capacity().value();
 		int numTrips = ceil((float)numPackages/(float)capacity);
-		forwardActivity_->nextTimeIs(activityManagerInstance()->now().value() + numTrips*length/speed);
+		forwardActivity_->nextTimeIs(activityManagerInstance()->now().value() + (float)numTrips*length/speed);
 		forwardActivity_->lastNotifieeIs(reactor_.ptr());
 		forwardActivity_->statusIs(Activity::queued);
   }
